Introduce functionality for automatically fetching package versions
Versions of the packages included in update file needs to be entered
manually in sw-description file or provided in as a bitbake variable.

This patch searches for a group[1] in sw-description file containing the
tag `"SWU_AUTO_VERSION"`.
If found, the version of the corresponding package is fetched from it's
package-data-file (`PV`) and the tag is replaced with the version.
Possible trailing strings in `PV` starting with `+` (e.g.
"+gitAUTOINC+...") are dropped to keep semantic versions.

Since the filenames not always belong to the package name (deployed
with another name or the file is a container for the real package)
support providing the correct package name by appending the tag with
`:<package-name>`.

[1]: http://www.hyperrealm.com/libconfig/libconfig_manual.html#Groups

Signed-off-by: Thomas Haemmerle <thomas.haemmerle@leica-geosystems.com>